Order düzenlemesi

+++ VkApi/Vk.Schema/Product.cs
@@ -5,7 +5,7 @@ public class CreateProductRequest
     public string Name           { get; set; }
     public string Category       { get; set; }
     public int    Stock          { get; set; }
-    public float  Price          { get; set; }
+    public decimal  Price          { get; set; }
     public string imgUrl       { get; set; }
 
 }
@@ -15,7 +15,7 @@ public class UpdateProductRequest
     public string Name           { get; set; }
     public string Category       { get; set; }
     public int    Stock          { get; set; }
-    public float  Price          { get; set; }
+    public decimal  Price          { get; set; }
     public string imgUrl       { get; set; }
 }
 
@@ -25,7 +25,7 @@ public class ProductResponse
     public string Name           { get; set; }
     public string Category       { get; set; }
     public int    Stock          { get; set; }
-    public float  Price          { get; set; }
+    public decimal  Price          { get; set; }
     public string imgUrl         { get; set; }
     public bool IsActive         { get; set; }
 }
